implement blueprints #27
