+++ app/chatbot/shop.py
