+++ app/chatbot/templates/favorites.json
