+++ app/chatbot/templates/favoritesComfirmMessage.json
