--- app/db.py
@@ -1,1 +1,0 @@
-from . import db
